refactor(closePayment): move `closePayment` business logic into separate helper class
